refactor: Player 이동 속도를 float에서 int로 변경

+++ Assets/02. Scipts/InGame/Choice/Choice13.cs
@@ -6,7 +6,7 @@ public class Choice13 : ChoiceBase
     private static readonly int ASpiritGainAmount = 1;
 
     private static readonly int BAdditionalDamageAmount = 5;
-    private static readonly float BAdditionalSpeed = 1f;
+    private static readonly int BAdditionalSpeed = 1;
     private static readonly int BAdditionalMaxHealth = 5;
 
     protected override void StepA()
